remove UI from javadoc and rename class name
add missing getMappingInfo() to each view in demo-web-site project
remove unnecessary servelet config.
