--- demo-jquery-table/src/main/java/com/vaadin/flow/demo/jquerytable/page/DemoUI.java
@@ -1,37 +1,0 @@
-/*
- * Copyright 2000-2017 Vaadin Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.vaadin.flow.demo.jquerytable.page;
-
-import javax.servlet.annotation.WebServlet;
-
-import com.vaadin.flow.component.UI;
-import com.vaadin.flow.server.VaadinServlet;
-import com.vaadin.flow.server.VaadinServletConfiguration;
-
-/**
- * The main UI for the application.
- */
-public class DemoUI extends UI {
-
-    /**
-     * The main servlet for the application.
-     */
-    @WebServlet(urlPatterns = "/*", name = "Servlet", asyncSupported = true)
-    @VaadinServletConfiguration(ui = DemoUI.class, productionMode = false)
-    public static class Servlet extends VaadinServlet {
-    }
-
-}
