+++ demo-web-site/src/main/java/com/vaadin/flow/demo/website/MainLayout.java
@@ -116,6 +116,5 @@ public final class MainLayout extends PolymerTemplate<MainLayoutModel>
 
         getModel().getItems().stream()
                 .forEach(item -> updateActive(item, navigatedToFirstSegment));
-        
     }
 }
